made limitation for max number length 12, fixed shared phonenumbers, now displays data correctly. Change number saving type from integer to bigInt

+++ database/migrations/2020_08_27_143951_create_phonenumbers_table.php
@@ -16,7 +16,7 @@ class CreatePhonenumbersTable extends Migration
         Schema::create('phonenumbers', function (Blueprint $table) {
             $table->id();
             $table->integer('user_id');
-            $table->integer('phonenumber'); 
+            $table->bigInteger('phonenumber'); 
         });
     }
